requirements.txt
python-dotenv==0.21.0
urllib3==1.26.9
